--- cmd/abigen/cmd_test.go
@@ -1,10 +1,0 @@
-package main
-
-import "testing"
-
-func TestGenerate(t *testing.T) {
-	err := Generate("example", "main", "token", "example/token.json", true)
-	if err != nil {
-		t.Error(err)
-	}
-}
